@@ -1,64 +1,81 @@
 {
   "name": "calendar-for-trello",
-  "version": "2.0.0-beta.23",
+  "version": "2.1.7",
   "license": "MIT",
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "hmr": "ng serve --hmr -e=hmr",
+    "hmr": "ng serve",
     "deploy": "ng github-pages:deploy",
-    "build": "ng build --prod --env=prod --aot",
+    "build": "run-p \"build:* -- --stats-json --configuration=production\"",
+    "build:en": "ng build",
+    "build:de": "ng build --base-href=/de/ --output-path=dist/de --i18n-file=src/locale/messages.de.xlf --i18n-format=xlf --i18n-locale=de",
+    "build:pt": "ng build --base-href=/pt/ --output-path=dist/pt --i18n-file=src/locale/messages.pt.xlf --i18n-format=xlf --i18n-locale=pt",
+    "build:fr": "ng build --base-href=/fr/ --output-path=dist/fr --i18n-file=src/locale/messages.fr.xlf --i18n-format=xlf --i18n-locale=fr",
+    "build:es": "ng build --base-href=/es/ --output-path=dist/es --i18n-file=src/locale/messages.es.xlf --i18n-format=xlf --i18n-locale=es",
+    "build:it": "ng build --base-href=/it/ --output-path=dist/it --i18n-file=src/locale/messages.it.xlf --i18n-format=xlf --i18n-locale=it",
     "start": "npm run hmr",
     "lint": "tslint \"src/**/*.ts\"",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor",
-    "i18n": "./node_modules/.bin/ng-xi18n -p ./src/tsconfig.json"
+    "extract-i18n": "ng xi18n --output-path locale && xliffmerge --profile xliffmerge.json en de pt fr es it"
   },
   "private": true,
   "dependencies": {
-    "@angular-redux/router": "^6.3.0",
-    "@angular-redux/store": "^6.4.5",
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "2.0.0-beta.7",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular-redux/router": "9.0.0",
+    "@angular-redux/store": "9.0.0",
+    "@angular/animations": "6.0.4",
+    "@angular/cdk": "6.2.1",
+    "@angular/common": "6.0.4",
+    "@angular/compiler": "6.0.4",
+    "@angular/core": "6.0.4",
+    "@angular/flex-layout": "^6.0.0-beta.16",
+    "@angular/forms": "6.0.4",
+    "@angular/http": "6.0.4",
+    "@angular/material": "^6.4.2",
+    "@angular/platform-browser": "6.0.4",
+    "@angular/platform-browser-dynamic": "6.0.4",
+    "@angular/router": "6.0.4",
     "@angularclass/hmr": "^1.2.2",
+    "@beyerleinf/ngx-dnd": "^6.0.0-preview3",
+    "@ngxs/devtools-plugin": "^3.2.0",
+    "@ngxs/storage-plugin": "^3.2.0",
+    "@ngxs/store": "^3.2.0",
+    "bootstrap": "^4.1.3",
     "classlist.js": "^1.1.20150312",
+    "cookieconsent": "3.0.6",
     "core-js": "^2.4.1",
+    "date-fns": "^1.29.0",
     "hammerjs": "^2.0.8",
+    "html2pdf.js": "^0.9.2",
     "intl": "^1.2.5",
-    "lodash": "^4.17.2",
+    "jspdf": "^1.5.3",
+    "lodash": "^4.17.10",
     "moment": "^2.15.1",
-    "ng2-dnd": "^2.1.1",
-    "ng2-rx-componentdestroyed": "^1.1.2",
+    "ng2-rx-componentdestroyed": "^3.0.0",
+    "node-sass": "^4.12.0",
     "raven-js": "^3.8.1",
-    "redux": "^3.6.0",
-    "redux-localstorage": "^0.4.0",
-    "redux-logger": "^2.6.1",
+    "redux": "^4.0.0",
+    "redux-localstorage": "^0.4.1",
+    "redux-logger": "^3.0.6",
     "reselect": "^3.0.1",
-    "rxjs": "^5.5.2",
-    "sass-loader": "^4.1.0",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.2.0",
+    "sass-loader": "^7.0.3",
+    "save": "^2.3.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.0-rc.1",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.4",
+    "@angular/language-service": "6.0.4",
     "@types/hammerjs": "^2.0.33",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/lodash": "^4.14.43",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^4.3.0",
+    "google-translate-xlf": "^0.2.3",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -67,9 +84,11 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "ngx-i18nsupport": "^0.17.1",
+    "npm-run-all": "^4.1.3",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "tslint": "~5.11.0",
+    "typescript": "2.7.2"
   }
 }
