@@ -3,15 +3,15 @@ import {CommonModule} from '@angular/common';
 import {AboutPageComponent} from './about-page/about-page.component';
 import {W11kInfoComponent} from './w11k-info/w11k-info.component';
 import {RouterModule} from '@angular/router';
-import {MdButtonModule, MdCardModule, MdIconModule} from "@angular/material";
+import {MatButtonModule, MatCardModule, MatIconModule} from '@angular/material';
 
 @NgModule({
   imports: [
     CommonModule,
     RouterModule,
-    MdCardModule,
-    MdButtonModule,
-    MdIconModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
   ],
   declarations: [AboutPageComponent, W11kInfoComponent],
   exports: [W11kInfoComponent],
