@@ -1,25 +1,22 @@
-import {Component, OnInit, HostListener, OnDestroy} from '@angular/core';
-import {RootState, enhancers} from './redux/store/index';
+import {Component, HostListener, OnDestroy, OnInit} from '@angular/core';
+import {enhancers, RootState} from './redux/store';
 import {NgReduxRouter} from '@angular-redux/router';
-import reducer from '../app/redux/reducers/index';
-import * as moment from 'moment';
-import {Router, NavigationEnd} from '@angular/router';
+import reducer from './redux/reducers';
+import {NavigationEnd, Router} from '@angular/router';
 import {NgRedux, select} from '@angular-redux/store';
 import {TrelloPullService} from './services/trello-pull.service';
 import {Settings} from './models/settings';
 import {Observable, Subscription} from 'rxjs';
 import {TrelloAuthService} from './services/trello-auth.service';
-import {MdSnackBar} from '@angular/material';
-import {IS_UPDATE} from '../main';
+import {MatSnackBar} from '@angular/material';
 import {SettingsActions} from './redux/actions/settings-actions';
 import {ListActions} from './redux/actions/list-actions';
 import {CardActions} from './redux/actions/card-actions';
 import {UserActions} from './redux/actions/user-actions';
 import {BoardActions} from './redux/actions/board-actions';
+
 const project = require('../../package.json');
 declare let ga: Function;
-
-// declare const IS_UPDATE:boolean;
 
 @Component({
   selector: 'app-root',
@@ -42,19 +39,24 @@ export class AppComponent implements OnInit, OnDestroy {
     },
     settings: new Settings(),
     lists: {},
-    members: {}
+    members: new Map(),
+    router: {}
   };
 
   public isSidenavOpen = false;
+  IS_UPDATE = false;
+  PROJECT_VERSION: string = project.version;
 
   constructor(private ngRedux: NgRedux<RootState>,
               private ngReduxRouter: NgReduxRouter,
               public router: Router,
               private trelloPullService: TrelloPullService,
               private trelloAuthService: TrelloAuthService,
-              private snackBar: MdSnackBar) {
+              private snackBar: MatSnackBar) {
 
-    if (IS_UPDATE) {
+    this.check();
+
+    if (this.IS_UPDATE) {
       this.snackBar.open('Calendar for Trello was updated to version ' + project.version + '!', 'OK');
     }
 
@@ -81,17 +83,21 @@ export class AppComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
     this.subscriptions.push(
-      this.settings$.subscribe(
-        settings => {
-          this.settings = settings;
-          // moment.locale(settings.language);
-        }
-      )
-    );
-
-    this.subscriptions.push(
       this.settings$.subscribe(settings => this.settings = settings)
     );
+
+    if (this.settings.weekViewShowHours === undefined) {
+      this.settings.weekViewShowHours = true;
+    }
+    if (this.settings.showWeekend === undefined) {
+      this.settings.showWeekend = true;
+    }
+    if (this.settings.businessHoursStart === undefined) {
+      this.settings.businessHoursStart = 9;
+    }
+    if (this.settings.businessHoursEnd === undefined) {
+      this.settings.businessHoursEnd = 18;
+    }
 
     this.checkWidth(window.innerWidth);
   }
@@ -121,9 +127,7 @@ export class AppComponent implements OnInit, OnDestroy {
     this.ngRedux.dispatch({type: CardActions.RESET_CARD_STORE});
     this.ngRedux.dispatch({type: ListActions.RESET_LIST_STORE});
     this.ngRedux.dispatch({type: SettingsActions.RESET_SETTINGS_STORE});
-    setTimeout(() => {
-      this.trelloPullService.pull();
-    }, 500);
+    this.logout();
   }
 
   toIssue() {
@@ -133,77 +137,24 @@ export class AppComponent implements OnInit, OnDestroy {
 
   trigger() {
 
-    let myform = document.createElement('form');
+    const myform = document.createElement('form');
     myform.action = 'https://www.paypal.com/cgi-bin/webscr';
     myform.method = 'post';
     myform.target = '_blank';
     myform.className = 'hidden';
 
-    let cmd = document.createElement('input');
+    const cmd = document.createElement('input');
     cmd.name = 'cmd';
     cmd.type = 'hidden';
-    cmd.value = '_donations';
+    cmd.value = '_s-xclick';
 
-    let encrypted = document.createElement('input');
-    encrypted.type = 'hidden';
-    encrypted.name = 'business';
-    encrypted.value = 'payments@w11k.de';
-
-    let lc = document.createElement('input');
-    lc.type = 'hidden';
-    lc.name = 'lc';
-    lc.value = 'GB';
-
-    let itemname = document.createElement('input');
-    itemname.type = 'hidden';
-    itemname.name = 'item_name';
-    itemname.value = 'Trello Calendar - W11K GmbH';
-
-    let itemnumber = document.createElement('input');
-    itemnumber.type = 'hidden';
-    itemnumber.name = 'item_number';
-    itemnumber.value = '2013';
-
-    let note = document.createElement('input');
-    note.type = 'hidden';
-    note.name = 'no_note';
-    note.value = '0';
-
-    let currency = document.createElement('input');
-    currency.type = 'hidden';
-    currency.name = 'currency_code';
-    currency.value = 'USD';
-
-    let bn = document.createElement('input');
-    bn.type = 'hidden';
-    bn.name = 'bn';
-    bn.value = 'PP-DonationsBF:btn_donate_LG.gif:NonHostedGuest';
-
-    let image = document.createElement('input');
-    image.type = 'image';
-    image.src = 'https://www.paypalobjects.com/en_US/i/btn/btn_donateCC_LG.gif';
-    image.border = '0';
-    image.id = 'Dimage';
-    image.name = 'submit';
-    image.alt = 'PayPal - The safer, easier way to pay online!';
-
-    //Most probably this can be skipped, but I left it in here since it was present in the generated code
-    let pixel: any = document.createElement('image');
-    pixel.border = '0';
-    pixel.alt = '';
-    pixel.src = 'https://www.paypalobjects.com/en_US/i/scr/pixel.gif';
-    pixel.width = '1';
-    pixel.height = '1';
+    const hosted = document.createElement('input');
+    hosted.name = 'hosted_button_id';
+    hosted.type = 'hidden';
+    hosted.value = 'E2NMK8X28A7PC';
 
     myform.appendChild(cmd);
-    myform.appendChild(encrypted);
-    myform.appendChild(lc);
-    myform.appendChild(itemname);
-    myform.appendChild(note);
-    myform.appendChild(currency);
-    myform.appendChild(bn);
-    myform.appendChild(image);
-    myform.appendChild(pixel);
+    myform.appendChild(hosted);
     document.body.appendChild(myform); // Not sure if this step is necessary
 
     myform.submit();
@@ -213,6 +164,39 @@ export class AppComponent implements OnInit, OnDestroy {
       myform.remove();
     }, 500);
 
+  }
+
+
+  check() {
+    let token = localStorage.getItem('trello_token') || localStorage.getItem('token');
+
+    if (!token) {
+      // no token, fresh user
+      return;
+    }
+    token = token.replace(/"/g, '');
+
+    const dataVersion = localStorage.getItem('version');
+
+    // let PROJECT_VERSION;
+    if (dataVersion !== this.PROJECT_VERSION) {
+      // older v2 version
+
+      // remove usage data (keeps settings!)
+      localStorage.removeItem('w11k.trello-cal/members');
+      localStorage.removeItem('w11k.trello-cal/boards');
+      localStorage.removeItem('w11k.trello-cal/cards');
+
+      this.IS_UPDATE = true;
+      localStorage.setItem('version', this.PROJECT_VERSION);
+    }
+
+    if (!dataVersion) {
+      // old v1 version, clear everything
+      localStorage.clear();
+      localStorage.setItem('token', token);
+      localStorage.setItem('version', this.PROJECT_VERSION);
+    }
   }
 
 }
