@@ -1,17 +1,19 @@
-import {Component, OnInit, Input, HostBinding, OnDestroy} from '@angular/core';
+import {ChangeDetectionStrategy, Component, HostBinding, Input, OnDestroy, OnInit} from '@angular/core';
 import {Card} from '../../models/card';
 import {select} from '@angular-redux/store';
-import {Observable, Subscription} from 'rxjs';
+import {combineLatest, Observable, Subscription} from 'rxjs';
 import {Board} from '../../models/board';
-import * as _ from 'lodash';
 import {List} from '../../models/list';
 import {selectBoardColorPrefs} from '../../redux/store/selects';
 import {MemberMap} from '../../redux/reducers/member.reducer';
+import {CardActions} from '../../redux/actions/card-actions';
+import {addMonths, addWeeks} from 'date-fns';
 
 @Component({
   selector: 'app-calendar-card',
   templateUrl: './calendar-card.component.html',
-  styleUrls: ['./calendar-card.component.scss']
+  styleUrls: ['./calendar-card.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class CalendarCardComponent implements OnInit, OnDestroy {
 
@@ -28,24 +30,24 @@ export class CalendarCardComponent implements OnInit, OnDestroy {
   @HostBinding('style.border-left-color') borderLeft;
   @HostBinding('class.dueComplete') dueComplete;
   @Input() public card: Card;
+  @Input() public showTime: Boolean = false;
 
-  constructor() {
+  constructor(private cardActions: CardActions) {
   }
 
   getAvatar(userId: string) {
-    return this.memberMap ? (this.memberMap[userId] ? this.memberMap[userId].avatarHash : '' ) : '';
+    return this.memberMap[userId].avatarUrl;
   }
 
   ngOnInit() {
     this.subscriptions.push(
-      Observable
-        .combineLatest(this.boardColorPrefs$, this.boards$, this.lists$)
+      combineLatest(this.boardColorPrefs$, this.boards$, this.lists$)
         .subscribe(x => {
           const boardColorPrefs = x[0];
           const boards: Board[] = x[1];
           const lists = x[2];
           this.list = lists ? lists[this.card.idList] : '';
-          this.board = _.find(boards, (board: Board) => board.id === this.card.idBoard);
+          this.board = boards.find((board: Board) => board.id === this.card.idBoard);
           this.borderLeft = boardColorPrefs[this.card.idBoard] || (this.board ? this.board.prefs.backgroundColor : null);
         }));
 
@@ -62,4 +64,11 @@ export class CalendarCardComponent implements OnInit, OnDestroy {
     this.subscriptions.forEach(subscription => subscription.unsubscribe());
   }
 
+  plus(amount: number, unit: 'week' | 'month') {
+    const nextDue = unit === 'month'
+      ? addMonths(this.card.due, amount)
+      : addWeeks(this.card.due, amount);
+
+    this.cardActions.updateCardsDue(this.card.id, nextDue);
+  }
 }
