@@ -1,5 +1,5 @@
-import {Component, OnInit, HostBinding, ViewChild} from '@angular/core';
-import {MdMenuTrigger, MdMenu} from '@angular/material';
+import {Component, HostBinding, OnInit, ViewChild} from '@angular/core';
+import {MatMenu, MatMenuTrigger} from '@angular/material';
 import {ContextMenuService} from './context-menu.service';
 
 @Component({
@@ -9,12 +9,12 @@ import {ContextMenuService} from './context-menu.service';
 })
 export class ContextMenuHolderComponent implements OnInit {
 
-  @ViewChild(MdMenuTrigger) trigger: MdMenuTrigger;
-  @ViewChild(MdMenu) menu: MdMenu;
+  @ViewChild(MatMenuTrigger) trigger: MatMenuTrigger;
+  @ViewChild(MatMenu) menu: MatMenu;
 
 
-  @HostBinding('style.top') top: string = '200px';
-  @HostBinding('style.left') left: string = '200px';
+  @HostBinding('style.top') top = '200px';
+  @HostBinding('style.left') left = '200px';
 
   constructor(private contextMenuService: ContextMenuService) {
     this.contextMenuService.register(this);
