@@ -1,6 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {TrelloAuthService} from '../services/trello-auth.service';
-import {MdDialog, MdDialogConfig} from '@angular/material';
+import {MatDialog, MatDialogConfig} from '@angular/material';
 import {DialogComponent} from './dialog/dialog.component';
 
 @Component({
@@ -10,13 +9,13 @@ import {DialogComponent} from './dialog/dialog.component';
 })
 export class FrontPageComponent implements OnInit {
 
-  constructor(public dialog: MdDialog) {
+  constructor(public dialog: MatDialog) {
 
-    let config = new MdDialogConfig();
+    const config = new MatDialogConfig();
     config.disableClose = true;
 
     // disableClose
-    let dialogRef = this.dialog.open(DialogComponent, config);
+    const dialogRef = this.dialog.open(DialogComponent, config);
     // dialogRef.afterClosed().subscribe(result => {
     //   // this.selectedOption = result;
     // });
