@@ -3,7 +3,7 @@ import {CommonModule} from '@angular/common';
 import {FrontPageComponent} from './front-page.component';
 import {DialogComponent} from './dialog/dialog.component';
 import {CalendarModule} from '../calendar/calendar.module';
-import {MdIconModule, MdToolbarModule} from '@angular/material';
+import {MatButtonModule, MatIconModule, MatToolbarModule} from '@angular/material';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {AboutModule} from '../about/about.module';
 import {RouterModule} from '@angular/router';
@@ -12,11 +12,12 @@ import {RouterModule} from '@angular/router';
   imports: [
     CommonModule,
     CalendarModule,
-    MdToolbarModule,
+    MatToolbarModule,
     FlexLayoutModule,
     AboutModule,
     RouterModule,
-    MdIconModule,
+    MatIconModule,
+    MatButtonModule,
   ],
   declarations: [
     FrontPageComponent,
