@@ -1,8 +1,11 @@
 export class CalendarDay {
+  date: Date;
 
-  constructor(public date: Date,
+  constructor(date: Date,
               public isDayOff?: boolean,
               public isToday?: boolean) {
+
+    this.date = new Date(date);
 
   }
 }
