@@ -1,8 +1,11 @@
+import {Board} from './board';
+
 export class Member {
-  avatarHash: string;
+  avatarUrl: string;
   constructor(public id?: string,
               public fullName?: string,
-              public username?: string) {
+              public username?: string,
+              public boards?: Board[]) {
 
   }
 }
