@@ -1,7 +1,8 @@
 export class MenuItem {
   constructor(public text: string,
               public link: string,
-              public icon: string) {
+              public icon: string,
+              public hot: boolean = false) {
 
   }
 }
