@@ -4,9 +4,9 @@ import {Injectable} from '@angular/core';
 import {NgRedux} from '@angular-redux/store';
 import {RootState} from '../store';
 import {CalendarDay} from '../../models/calendar-day';
-import {CalendarService} from '../../services/calendar.service';
-import {Moment} from 'moment';
+import {CalendarDaysService} from '../../services/calendar-days.service';
 import {CalendarType} from './settings-actions';
+import {addMonths, addWeeks, subMonths, subWeeks} from 'date-fns';
 
 export enum PeriodChange {
   add, subtract
@@ -14,57 +14,51 @@ export enum PeriodChange {
 
 @Injectable()
 export class CalendarActions {
-  static FILL_DAYS: string = 'FILL_DAYS';
-  static BUILD_DAYS: string = 'BUILD_DAYS';
-  static NAVIGATE: string = 'NAVIGATE';
-  static RESET_CALENDAR_STORE: string = 'RESET_CALENDAR_STORE';
+  static BUILD_DAYS = 'BUILD_DAYS';
+  static RESET_CALENDAR_STORE = 'RESET_CALENDAR_STORE';
 
   constructor(private ngRedux: NgRedux<RootState>,
-              private calendarService: CalendarService) {
+              private calendarService: CalendarDaysService) {
   }
 
-  public fillDays(days: CalendarDay) {
-    this.ngRedux.dispatch({type: CalendarActions.FILL_DAYS, payload: days});
-  };
-
-  public buildDays(date: Moment, calendarType: CalendarType = CalendarType.Month) {
+  public buildDays(date: Date, calendarType: CalendarType = CalendarType.Month) {
     this.calendarService.buildDaysAsync(date, calendarType)
       .then(days => {
         this.ngRedux.dispatch({type: CalendarActions.BUILD_DAYS, payload: days, date: date});
       })
     ;
-  };
+  }
 
 
-
-  public navigate(date: Moment, change: PeriodChange, calendarType: CalendarType) {
+  public navigate(date: Date, change: PeriodChange, calendarType: CalendarType) {
     switch (calendarType) {
       case CalendarType.Month:
         switch (change) {
           case PeriodChange.add:
-            date.add(1, 'months');
+            date = addMonths(date, 1);
             break;
           case PeriodChange.subtract:
-            date.subtract(1, 'months');
+            date = subMonths(date, 1);
             break;
         }
         break;
       case CalendarType.Week:
+      case CalendarType.WorkWeek:
         switch (change) {
           case PeriodChange.add:
-            date.add(1, 'weeks');
+            date = addWeeks(date, 1);
             break;
           case PeriodChange.subtract:
-            date.subtract(1, 'weeks');
+            date = subWeeks(date, 1);
             break;
         }
         break;
     }
     this.buildDays(date, calendarType);
-  };
+  }
 
   // navigates to a date
-  public navigateToDate(date: Moment, calendarType: CalendarType) {
+  public navigateToDate(date: Date, calendarType: CalendarType) {
     this.buildDays(date, calendarType);
 
   }
