@@ -8,42 +8,43 @@ import {BoardActions} from './board-actions';
 @Injectable()
 export class CardActions {
 
-  static ADD_CARD: string = 'ADD_CARD';
-  static UPDATE_CARD: string = 'UPDATE_CARD';
-  static UPDATE_CARDS_OF_BOARD: string = 'UPDATE_CARDS_OF_BOARD';
-  static REBUILD_STORE: string = 'REBUILD_STORE';
-  static UPDATE_DUE: string = 'UPDATE_DUE';
-  static REMOVE_DUE: string = 'REMOVE_DUE';
-  static ARCHIVE_CARD: string = 'ARCHIVE_CARD';
-  static MARK_CARD_DONE: string = 'MARK_CARD_DONE';
-  static REMOVE_CARDS_BY_BOARDID: string = 'REMOVE_CARDS_BY_BOARDID';
-  static RESET_CARD_STORE: string = 'RESET_CARD_STORE';
+  static ADD_CARD = 'ADD_CARD';
+  static UPDATE_CARD = 'UPDATE_CARD';
+  static UPDATE_CARDS_OF_BOARD = 'UPDATE_CARDS_OF_BOARD';
+  static REBUILD_STORE = 'REBUILD_STORE';
+  static UPDATE_DUE = 'UPDATE_DUE';
+  static REMOVE_DUE = 'REMOVE_DUE';
+  static ARCHIVE_CARD = 'ARCHIVE_CARD';
+  static MARK_CARD_DONE = 'MARK_CARD_DONE';
+  static REMOVE_CARDS_BY_BOARDID = 'REMOVE_CARDS_BY_BOARDID';
+  static RESET_CARD_STORE = 'RESET_CARD_STORE';
+  static CARD_CHANGE_LIST = 'CARD_CHANGE_LIST';
 
   constructor(private ngRedux: NgRedux<RootState>, private tHttp: TrelloHttpService) {
   }
 
   public addCard(card: Card) {
     this.ngRedux.dispatch({type: CardActions.ADD_CARD, payload: card});
-  };
+  }
 
   // inserts new cards from API
   public rebuildStore(cards: Card[]) {
     this.ngRedux.dispatch({type: CardActions.REBUILD_STORE, payload: cards});
-  };
+  }
 
   // i think this is bad:
   public updateCard(card: Card) {
     this.ngRedux.dispatch({type: CardActions.UPDATE_CARD, payload: card});
-  };
+  }
 
   public updateCardsDue(cardId: string, due: Date) {
     this.tHttp.put('cards/' + cardId + '', {
       due: due
     }).subscribe(
       success => this.ngRedux.dispatch({type: CardActions.UPDATE_DUE, id: cardId, due: due}),
-      error => console.log(error)
+      error => console.error(error)
     );
-  };
+  }
 
   /**
    * Only updates cards of a single board after loading*/
@@ -67,14 +68,14 @@ export class CardActions {
 
   public removeCardsByBoardId(boardId: string) {
     this.ngRedux.dispatch({type: CardActions.REMOVE_CARDS_BY_BOARDID, boardId: boardId});
-  };
+  }
 
   public removeDue(cardId: string) {
     this.tHttp.put('cards/' + cardId + '', {
       due: 'null'
     }).subscribe(
       success => this.ngRedux.dispatch({type: CardActions.REMOVE_DUE, id: cardId}),
-      error => console.log(error)
+      error => console.error(error)
     );
   }
 
@@ -83,7 +84,7 @@ export class CardActions {
       closed: true
     }).subscribe(
       success => this.ngRedux.dispatch({type: CardActions.ARCHIVE_CARD, id: cardId}),
-      error => console.log(error)
+      error => console.error(error)
     );
 
   }
@@ -97,10 +98,22 @@ export class CardActions {
       dueComplete: !card.dueComplete
     }).subscribe(
       success => this.ngRedux.dispatch({type: CardActions.MARK_CARD_DONE, payload: card}),
-      error => console.log(error)
+      error => console.error(error)
     );
 
 
+  }
+
+  //  Feature change card list - 05.06.18
+  public changerCardList(cardId: string, idBoard: string, idList: string) {
+    this.tHttp.put('cards/' + cardId + '', {
+      idList: idList,
+      cardId: cardId,
+      idBoard: idBoard,
+    }).subscribe(
+      success => this.ngRedux.dispatch({type: CardActions.CARD_CHANGE_LIST, id: cardId, idList: idList}),
+      error => console.error(error)
+    );
   }
 
 }
