@@ -5,25 +5,25 @@ import {User} from '../../models/user';
 
 @Injectable()
 export class UserActions {
-  static ADD_USER: string = 'ADD_USER';
-  static UPDATE_USER: string = 'UPDATE_USER';
-  static REMOVE_USER: string = 'REMOVE_USER';
-  static RESET_USER_STORE: string = 'RESET_USER_STORE';
+  static ADD_USER = 'ADD_USER';
+  static UPDATE_USER = 'UPDATE_USER';
+  static REMOVE_USER = 'REMOVE_USER';
+  static RESET_USER_STORE = 'RESET_USER_STORE';
 
   constructor(private ngRedux: NgRedux<RootState>) {
   }
 
   public addUser(user: User) {
     this.ngRedux.dispatch({type: UserActions.ADD_USER, payload: user});
-  };
+  }
 
   public updateUser(user: User) {
     this.ngRedux.dispatch({type: UserActions.UPDATE_USER, payload: user});
-  };
+  }
 
   public removeUser() {
     this.ngRedux.dispatch({type: UserActions.REMOVE_USER});
-  };
+  }
 
 
   public resetStore() {
