@@ -1,15 +1,14 @@
 // -> Takes previous state + action, returnes new // new state
 import {CalendarActions} from '../actions/calendar-actions';
 import {CalendarDay} from '../../models/calendar-day';
-import {Moment} from 'moment';
-import * as moment from 'moment';
 import {CalendarType} from '../actions/settings-actions';
 
 export interface CalendarState {
   days: CalendarDay[];
-  date: Moment;
+  date: Date;
 }
-const initialState = {days: [], date: moment(), type: CalendarType.Month};
+
+const initialState = {days: [], date: new Date(), type: CalendarType.Month};
 
 
 export default (state: CalendarState = initialState, action: any) => {
