@@ -3,16 +3,36 @@ import {MemberActions} from '../actions/member-actions';
 import {Member} from '../../models/member';
 import {Action} from './action';
 
+/*
 export interface MemberMap { [s: string]: Member;
 }
+*/
 
-const initialState: MemberMap = {};
+export type MemberMap = Map<string, Member>;
 
-export default (state: MemberMap = initialState, action: Action<Member>) => {
+const initialState: MemberMap = new Map();
+
+export default (state: MemberMap = initialState, action: any) => {
   switch (action.type) {
     case MemberActions.SET_MEMBER:
-      let memberObj = {};
-      memberObj[action.payload.id] = action.payload;
+      const memberFromState = state[action.member.id];
+      let boards;
+
+
+      if (!memberFromState) {
+        boards = [action.board];
+      } else {
+        const boardsOfMember = memberFromState.boards ? memberFromState.boards : [];
+        boardsOfMember.push(action.board);
+        boards = boardsOfMember;
+      }
+
+      const member: Member = Object.assign({}, action.member);
+      member.boards = boards;
+
+      const memberObj = {};
+      memberObj[action.member.id] = member;
+
       return Object.assign(state, memberObj);
     case MemberActions.RESET_MEMBER_STORE:
       return initialState;
