@@ -4,6 +4,7 @@ import {Board} from '../../models/board';
 import {CalendarState} from '../reducers/calendar.reducer';
 import {Settings} from '../../models/settings';
 import {Member} from '../../models/member';
+import {MemberMap} from '../reducers/member.reducer';
 
 // switch with https://www.npmjs.com/package/redux-replicate-localforage? Since localstorage is sync.
 const persistState = require('redux-localstorage');
@@ -20,13 +21,23 @@ export const enhancers = [
 if (window.devToolsExtension) {
   enhancers.push(window.devToolsExtension());
 }
+/*export type State = {
+  readonly cards: Card[];
+  readonly boards: Board[];
+  readonly user: User;
+  readonly calendar: CalendarState;
+  readonly settings: Settings;
+  readonly lists: Object;
+  readonly members: Object;
+}*/
 
 export interface RootState {
+  router: any;
   cards: Card[];
   boards: Board[];
   user: User;
   calendar: CalendarState;
   settings: Settings;
   lists: Object;
-  members: Object;
+  members: MemberMap;
 }
