@@ -3,20 +3,29 @@ import {CommonModule} from '@angular/common';
 import {SettingsComponent} from './settings.component';
 import {EditBoardComponent} from './edit-board/edit-board.component';
 import {FormsModule} from '@angular/forms';
-import {MdListModule, MdSelectModule, MdCardModule, MdSlideToggleModule, MdRadioModule, MdIconModule} from '@angular/material';
+import {
+  MatCardModule,
+  MatCheckboxModule,
+  MatIconModule,
+  MatListModule,
+  MatRadioModule,
+  MatSelectModule,
+  MatSlideToggleModule
+} from '@angular/material';
 import {FlexLayoutModule} from '@angular/flex-layout';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
-    MdListModule,
-    MdSelectModule,
-    MdRadioModule,
+    MatListModule,
+    MatSelectModule,
+    MatRadioModule,
     FlexLayoutModule,
-    MdCardModule,
-    MdSlideToggleModule,
-    MdIconModule,
+    MatCardModule,
+    MatSlideToggleModule,
+    MatIconModule,
+    MatCheckboxModule,
   ],
   declarations: [
     SettingsComponent,
