@@ -1,11 +1,8 @@
-/* tslint:disable:no-unused-variable */
-
-import {TestBed, async} from '@angular/core/testing';
 import {CutStringPipe} from './cut-string.pipe';
 
 describe('CutStringPipe', () => {
   it('create an instance', () => {
-    let pipe = new CutStringPipe();
+    const pipe = new CutStringPipe();
     expect(pipe).toBeTruthy();
   });
 });
