@@ -1,4 +1,4 @@
-import {NgModuleRef, ApplicationRef} from '@angular/core';
+import {ApplicationRef, NgModuleRef} from '@angular/core';
 import {createNewHosts} from '@angularclass/hmr';
 
 export const hmrBootstrap = (module: any, bootstrap: () => Promise<NgModuleRef<any>>) => {
@@ -6,9 +6,9 @@ export const hmrBootstrap = (module: any, bootstrap: () => Promise<NgModuleRef<a
   module.hot.accept();
   bootstrap().then(mod => ngModule = mod);
   module.hot.dispose(() => {
-    let appRef: ApplicationRef = ngModule.injector.get(ApplicationRef);
-    let elements = appRef.components.map(c => c.location.nativeElement);
-    let makeVisible = createNewHosts(elements);
+    const appRef: ApplicationRef = ngModule.injector.get(ApplicationRef);
+    const elements = appRef.components.map(c => c.location.nativeElement);
+    const makeVisible = createNewHosts(elements);
     ngModule.destroy();
     makeVisible();
   });
